refactor(treeshake): split include_statements.rs into focused modules (#10095)
### Description

Part 1 of the tree-shaking inclusion refactor stack.

`include_statements.rs` had grown to ~1200 lines interleaving the inclusion fixpoint driver, the three mutually recursive inclusion functions, dynamic-entry topological ordering, the body-demand gating model, and several standalone passes — with 10+ orthogonal policies (CJS bailouts, constant/enum inlining bypasses, JSON self-reference tracking, `eval` handling, HMR namespace registration, property-write retention) inlined directly into the recursion. Reviewing any change meant re-deriving which clause belonged to which concern; every bug fix added another inline branch.

Pure code motion, in three commits:

- **Split into focused modules**
  - `dynamic_entries.rs` — dynamic-entry topological ordering (tarjan), aliveness determination, per-iteration processing (~200 lines unrelated to inclusion mechanics)
  - `on_demand.rs` — the body-demand gating model (`side_effects_included_on_demand`, `compute_on_demand_side_effect_stmts`, `is_gated_side_effect_stmt`)
  - `passes.rs` — standalone passes around the fixpoint (CJS bailout export inclusion, runtime-helper collection/inclusion, `preserveModules` re-export interface preservation)
- **Extract named policies** — each inline block of `include_module` / `include_symbol` / `include_statement` becomes a named function with a doc contract (`sweep_side_effect_statements`, `include_side_effectful_dependencies`, `is_bypassed_inlined_constant`, `drain_body_demand_stmts`, `follow_cjs_namespace_alias`, `note_namespace_inclusion_reason`, `demand_esm_init_wrapper`, `note_json_self_reference`, `include_property_write_referencing_stmts`, `scan_import_records_for_cjs_bailout`, `is_inlined_enum_member_access`), so the three skeletons read as a sequence of policies. Also dedupes `include_declaring_statements`, which replaces five hand-rolled copies of the "a used binding keeps its declaring statements" loop.
- **wasm build fix** — the move dropped an incidental `itertools::Itertools` import that wasm targets need for `zip_eq` (`rolldown_utils::rayon` shims `par_iter()` to a plain `Iterator` there); caught by the browser/WASI CI jobs.

No behavior change: byte-identical output across the integration suite (zero snapshot diffs), clippy clean, `wasm32-wasip1-threads` checked.

Stack: #10095 → #10096 → #10097 → #10098
