+++ crates/rolldown/src/stages/link_stage/mod.rs
@@ -40,11 +40,8 @@ mod sort_modules;
 mod tree_shaking;
 
 pub use tree_shaking::{
-  ModuleInclusionVec, ModuleNamespaceReasonVec, StmtInclusionVec,
-  include_statements::{
-    IncludeContext, SymbolIncludeReason, compute_on_demand_side_effect_stmts,
-    include_runtime_symbol, include_symbol,
-  },
+  IncludeContext, ModuleInclusionVec, ModuleNamespaceReasonVec, StmtInclusionVec,
+  SymbolIncludeReason, compute_on_demand_side_effect_stmts, include_runtime_symbol, include_symbol,
 };
 mod wrapping;
